backend/requirements.txt
Flask
pymongo
python-dotenv
flask-cors
requests
pytest
Flask-Testing
requests-mock